move context location, add interface file, move BoredAPI to page.tsx instead
