+++ components/Interface.ts
@@ -1,0 +1,10 @@
+
+interface LocalDataInterface{
+    localData: {},
+    setLocalData: () => {}
+}
+
+interface BoredAPIResultInterface{
+    boredAPIData: null,
+    setBoredAPIData: () => {}
+}
